refactor(runtime): extract channel_send tool module
Largest single-tool extraction so far: tool_channel_send plus its two
private helpers parse_poll_options and mirror_channel_send_to_session
(issue #4824 mirror-back-to-session contract) all move into
tool_runner/channel.rs.

The channel tool handles four delivery modes — text, image_url /
file_url URL forwarding, local-file file_path attachment, and
Telegram-style polls — each branch verbatim. Taint checks
(check_taint_outbound_text) and the mirror-back call follow the
original ordering. parse_poll_options keeps fail-fast behavior on
non-string entries; mirror_channel_send_to_session keeps the JSON
envelope shape required for the inbound-routing session context.

To unblock the local-file branch, resolve_file_path_ext is promoted
to pub(super); the same change unblocks the upcoming image / media
extraction batch, which also reads files from the agent's workspace
sandbox.

mod.rs: 11,449 -> 11,058 LOC (-391). cargo check + cargo clippy --lib
both clean.

Refs #3710
